+++ src/LandingPage/LandingPage.jsx
@@ -10,7 +10,7 @@ export default class LandingPage extends Component {
     return(
       <div className="landingPage">
       <Nav />
-      <p><Link style={{textDecoration: 'none'}} to="Signup_Login">SignUp/Login</Link></p>
+      <p><Link style={{textDecoration: 'none'}} to="signup_login">SignUp/Login</Link></p>
       <Search />
       <div className="event_list">
