+++ src/components/Footer/Footer.css
@@ -3,10 +3,13 @@ position: absolute;
 bottom: -80px;
 width: 100%;
 display: block;
+margin-right: 0;
 margin-top: 50px;
 text-align: center;
 background-color: grey;
 }
 .footer p {
   text-align: center;
+  margin-right: 0;
+
 }
